move color-related code to `mod color`
