+++ node-graph/gcore/src/color/discrete_srgb.rs
